+++ apps/client/src/styles/GridStyle.tsx
